move validator function
