+++ api/app/routers/actionlogs.py
@@ -6,10 +6,8 @@ from sqlalchemy.orm import Session
 
 from app import models, persistence, schemas
 from app.auth import get_current_user
-from app.common import (
-    check_pteam_membership,
-)
 from app.database import get_db
+from app.validators.account_validator import check_pteam_membership
 
 router = APIRouter(prefix="/actionlogs", tags=["actionlogs"])
