+++ api/app/routers/actions.py
@@ -7,10 +7,8 @@ from sqlalchemy.orm import Session
 
 from app import models, persistence, schemas
 from app.auth import get_current_user
-from app.common import (
-    check_topic_action_tags_integrity,
-    fix_threats_for_topic,
-)
+from app.business.tag_business import check_topic_action_tags_integrity
+from app.common import fix_threats_for_topic
 from app.database import get_db
 
 router = APIRouter(prefix="/actions", tags=["actions"])
