+++ api/app/routers/ateams.py
@@ -7,14 +7,14 @@ from sqlalchemy.orm import Session
 
 from app import command, models, persistence, schemas
 from app.auth import get_current_user
-from app.common import (
+from app.constants import MEMBER_UUID, NOT_MEMBER_UUID, SYSTEM_UUID
+from app.database import get_db
+from app.slack import validate_slack_webhook_url
+from app.validators.account_validator import (
     check_ateam_auth,
     check_ateam_membership,
     check_pteam_auth,
 )
-from app.constants import MEMBER_UUID, NOT_MEMBER_UUID, SYSTEM_UUID
-from app.database import get_db
-from app.slack import validate_slack_webhook_url
 
 router = APIRouter(prefix="/ateams", tags=["ateams"])
